@@ -1,28 +1,28 @@
 name: retrofit_example
-description: Retrofit generator
-version: 0.0.1
-
+description: retrofit_example
+version: 1.0.0+1
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
+  dart_json_mapper: ^2.2.13
+  dio: ^5.7.0
+  http_parser: ^4.0.2
+  json_annotation: ^4.9.0
+  logger: ^2.4.0
+  reflectable: ^4.0.6
   retrofit:
-  json_annotation: ^4.0.0
-  logger: ^1.0.0
-  dart_json_mapper: ^2.0.5
-  reflectable: ^3.0.0
 
 dev_dependencies:
-  test: 1.16.5
+  build_runner: ^2.4.0
+  json_serializable: ^6.8.0
+  lints: ^4.0.0
+  mock_web_server: ^5.0.0-nullsafety.1
   retrofit_generator:
-  build_runner: ^2.0.1
-  json_serializable: ^4.0.3
-  mock_web_server:
+  test: ^1.25.0
 
 dependency_overrides:
   retrofit:
     path: ../retrofit
   retrofit_generator:
     path: ../generator
-  analyzer: ^1.1.0
-  dart_style: ^2.0.0
