@@ -3,26 +3,27 @@ description: retrofit generator is an dio client generator using source_gen and 
 homepage: https://mings.in/retrofit.dart/
 repository: https://github.com/trevorwang/retrofit.dart/
 topics:
-  - build-runner
-  - codegen
   - api
-version: 8.1.0
+  - http
+  - rest
+  - retrofit
+  - codegen
+version: 9.1.5
 environment:
-  sdk: '>=2.19.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
-  analyzer: '>=5.13.0 <7.0.0'
-  build: ^2.3.1
+  analyzer: ^6.5.0
+  build: ^2.4.1
   built_collection: ^5.1.1
-  code_builder: ^4.4.0
+  code_builder: ^4.10.0
   dart_style: ^2.3.0
   dio: ^5.0.0
-  retrofit: ^4.2.0
-  source_gen: ^1.3.0
-  tuple: ^2.0.1
   protobuf: ^3.1.0
+  retrofit: ^4.4.0
+  source_gen: ^1.5.0
 
 dev_dependencies:
-  lints: ^2.0.1
-  source_gen_test: ^1.0.5
-  test: ^1.24.0
+  lints: ^4.0.0
+  source_gen_test: ^1.0.6
+  test: ^1.25.0
