@@ -2,14 +2,19 @@ name: retrofit
 description: retrofit.dart is an dio client generator using source_gen and inspired by Chopper and Retrofit.
 homepage: https://mings.in/retrofit.dart/
 repository: https://github.com/trevorwang/retrofit.dart/
-version: 4.1.0
-
+topics:
+  - api
+  - http
+  - rest
+  - dio
+  - retrofit
+version: 4.4.0
 environment:
-  sdk: '>=2.17.0 <4.0.0'
+  sdk: '>=2.19.0 <4.0.0'
 
 dependencies:
   dio: ^5.0.0
   meta: ^1.8.0
 
 dev_dependencies:
-  lints: ^2.0.1
+  lints: ^4.0.0
