@@ -1,56 +1,56 @@
 import 'dart:io';
 
-import 'package:logger/logger.dart';
 import 'package:dio/dio.dart';
+import 'package:logger/logger.dart';
 import 'package:retrofit_example/example.dart';
 import 'package:retrofit_example/json_mapper_example.dart' hide Task;
 import 'package:retrofit_example/json_mapper_example.reflectable.dart'
     show initializeReflectable;
 
 final logger = Logger();
-void main(List<String> args) {
+
+Future<void> main(List<String> args) async {
   final dio = Dio();
-  dio.options.headers["Demo-Header"] = "demo header";
-  dio.options.headers["Content-Type"] = "application/json";
+  dio.options.headers['Demo-Header'] = 'demo header';
+  dio.options.headers['Content-Type'] = 'application/json';
   final client = RestClient(dio);
 
   client.getTasks().then((it) => logger.i(it));
 
-  client.getTask("2").then((it) => logger.i(it)).catchError((Object obj) {
+  client.getTask('2').then((it) => logger.i(it)).catchError((Object obj) {
     // non-200 error goes here.
     switch (obj.runtimeType) {
-      case DioError:
-        final res = (obj as DioError).response;
-        logger.e("Got error : ${res?.statusCode} -> ${res?.statusMessage}");
-        break;
+      case DioException _:
+        final res = (obj as DioException).response;
+        logger.e('Got error : ${res?.statusCode} -> ${res?.statusMessage}');
       default:
     }
   });
 
-  client.createTask(Task(avatar: "2222.png", name: "new task")).then((it) {
+  client.createTask(Task(avatar: '2222.png', name: 'new task')).then((it) {
     logger.i(it.toJson());
   });
 
   client
-      .updateTask("3", Task(id: "4", avatar: "1.png", name: "number 3"))
+      .updateTask('3', Task(id: '4', avatar: '1.png', name: 'number 3'))
       .then((it) {
     logger.i(it.toJson());
   });
 
   client
       .updateTaskPart(
-          "4", Task(id: "4", avatar: "1.png", name: "number 4").toJson())
+          '4', Task(id: '4', avatar: '1.png', name: 'number 4').toJson())
       .then((it) {
     logger.i(it.toJson());
   });
 
-  client.deleteTask("2").then((it) {
-    logger.i("taks 2 has been deleted!");
+  client.deleteTask('2').then((it) {
+    logger.i('taks 2 has been deleted!');
   }).catchError((Object err) {
     logger.e(err);
   });
 
-  client.createNewTaskFromFile(File("main.dart")).catchError((Object err) {
+  client.createNewTaskFromFile(File('main.dart')).catchError((Object err) {
     print(err);
   });
   client.getFile().then((it) {
@@ -59,5 +59,5 @@ void main(List<String> args) {
 
   initializeReflectable();
   final api = ApiService(dio);
-  api.getTasks(new DateTime.now()).then((it) => logger.i(it.toString()));
+  api.getTasks(DateTime.now()).then((it) => logger.i(it.toString()));
 }
