@@ -1,5 +1,4 @@
 import 'package:dio/dio.dart' hide Headers;
-import 'package:flutter/foundation.dart';
 import 'package:flutter_example/mock_adapter.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:retrofit/retrofit.dart';
@@ -8,13 +7,11 @@ part 'example.g.dart';
 
 User deserializeUser(Map<String, dynamic> json) => User.fromJson(json);
 
-List<User> deserializeUserList(List<Map<String, dynamic>> json) =>
-    json.map(User.fromJson).toList();
+List<User> deserializeUserList(List<Map<String, dynamic>> json) => json.map(User.fromJson).toList();
 
 Map<String, dynamic> serializeUser(User object) => object.toJson();
 
-List<Map<String, dynamic>> serializeUserList(List<User> objects) =>
-    objects.map((e) => e.toJson()).toList();
+List<Map<String, dynamic>> serializeUserList(List<User> objects) => objects.map((e) => e.toJson()).toList();
 
 @JsonSerializable()
 class User {
@@ -29,7 +26,7 @@ class User {
 
 @RestApi(baseUrl: 'http://baidu.com', parser: Parser.FlutterCompute)
 abstract class RestClient {
-  factory RestClient(Dio dio, {String baseUrl}) = _RestClient;
+  factory RestClient(Dio dio, {String? baseUrl}) = _RestClient;
 
   @GET('/tags')
   Future<List<String>> getTags({@DioOptions() Options? options});
@@ -110,10 +107,10 @@ abstract class RestClient {
   });
 }
 
-void test() {
-  final dio = Dio();
-  // remove this line when you want to test with real server
-  dio.httpClientAdapter = MockAdapter();
+Future<void> test() async {
+  final dio = Dio()
+    // remove this line when you want to test with real server
+    ..httpClientAdapter = MockAdapter();
 
   dio.interceptors.add(
     InterceptorsWrapper(
@@ -127,7 +124,5 @@ void test() {
     ),
   );
   final api = RestClient(dio, baseUrl: MockAdapter.mockBase);
-  api.getUsers().then((it) {
-    print(it.length);
-  });
+  await api.getUsers().then((it) => print(it.length));
 }
