@@ -13,10 +13,9 @@ import 'package:dio/dio.dart';
 import 'package:protobuf/protobuf.dart';
 import 'package:retrofit/retrofit.dart' as retrofit;
 import 'package:source_gen/source_gen.dart';
-import 'package:tuple/tuple.dart';
 
 const _analyzerIgnores =
-    '// ignore_for_file: unnecessary_brace_in_string_interps,no_leading_underscores_for_local_identifiers,unused_element';
+    '// ignore_for_file: unnecessary_brace_in_string_interps,no_leading_underscores_for_local_identifiers,unused_element,unnecessary_string_interpolations';
 
 class RetrofitOptions {
   RetrofitOptions({
@@ -226,7 +225,7 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
 
   Iterable<Method> _parseMethods(ClassElement element) => <MethodElement>[
         ...element.methods,
-        ...element.mixins.expand((i) => i.methods)
+        ...element.mixins.expand((i) => i.methods),
       ].where((m) {
         final methodAnnotation = _getMethodAnnotation(m);
         return methodAnnotation != null &&
@@ -319,14 +318,14 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
     return annotation;
   }
 
-  Tuple2<ParameterElement, ConstantReader>? _getAnnotation(
+  ({ParameterElement element, ConstantReader reader})? _getAnnotation(
     MethodElement m,
     Type type,
   ) {
     for (final p in m.parameters) {
       final a = _typeChecker(type).firstAnnotationOf(p);
       if (a != null) {
-        return Tuple2(p, ConstantReader(a));
+        return (element: p, reader: ConstantReader(a));
       }
     }
     return null;
@@ -387,7 +386,7 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
             : MethodModifier.asyncStar
         ..annotations.add(const CodeExpression(Code('override')));
 
-      if (globalOptions.useResult == true) {
+      if (globalOptions.useResult ?? false) {
         final returnType = m.returnType;
         if (returnType is ParameterizedType &&
             returnType.typeArguments.first is! VoidType) {
@@ -402,8 +401,7 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
                 (p) => p
                   ..name = it.name
                   ..named = it.isNamed
-                  ..type =
-                      refer(it.type.getDisplayString(withNullability: true)),
+                  ..type = refer(it.type.getDisplayString()),
               ),
             ),
       );
@@ -418,8 +416,7 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
                       !it.hasDefaultValue)
                   ..name = it.name
                   ..named = it.isNamed
-                  ..type =
-                      refer(it.type.getDisplayString(withNullability: true))
+                  ..type = refer(it.type.getDisplayString())
                   ..defaultTo = it.defaultValueCode == null
                       ? null
                       : Code(it.defaultValueCode!),
@@ -499,12 +496,13 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
 
     /// gen code for request body for content-type on Protobuf body
     final annotation = _getAnnotation(m, retrofit.Body);
-    final bodyName = annotation?.item1;
+    final bodyName = annotation?.element;
     if (bodyName != null) {
       if (const TypeChecker.fromRuntime(GeneratedMessage)
           .isAssignableFromType(bodyName.type)) {
         extraOptions[_contentType] = literal(
-            "application/x-protobuf; \${${bodyName.displayName}.info_.qualifiedMessageName == \"\" ? \"\" :\"messageType=\${${bodyName.displayName}.info_.qualifiedMessageName}\"}");
+          'application/x-protobuf; \${${bodyName.displayName}.info_.qualifiedMessageName == "" ? "" :"messageType=\${${bodyName.displayName}.info_.qualifiedMessageName}"}',
+        );
       }
     }
 
@@ -537,25 +535,27 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
 
     final cancelToken = _getAnnotation(m, retrofit.CancelRequest);
     if (cancelToken != null) {
-      namedArguments[_cancelToken] = refer(cancelToken.item1.displayName);
+      namedArguments[_cancelToken] = refer(cancelToken.element.displayName);
     }
 
     final sendProgress = _getAnnotation(m, retrofit.SendProgress);
     if (sendProgress != null) {
-      namedArguments[_onSendProgress] = refer(sendProgress.item1.displayName);
+      namedArguments[_onSendProgress] = refer(sendProgress.element.displayName);
     }
 
     final receiveProgress = _getAnnotation(m, retrofit.ReceiveProgress);
     if (receiveProgress != null) {
       namedArguments[_onReceiveProgress] =
-          refer(receiveProgress.item1.displayName);
+          refer(receiveProgress.element.displayName);
     }
 
     final wrappedReturnType = _getResponseType(m.returnType);
 
-    blocks.add(declareFinal(_optionsVar)
-        .assign(_parseOptions(m, namedArguments, blocks, extraOptions))
-        .statement);
+    blocks.add(
+      declareFinal(_optionsVar)
+          .assign(_parseOptions(m, namedArguments, blocks, extraOptions))
+          .statement,
+    );
 
     final options = refer(_optionsVar).expression;
 
@@ -620,7 +620,7 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
                         innerReturnType,
                         withNullability: innerReturnType?.isNullable ?? false,
                       ),
-                    )
+                    ),
                   ]),
                 )
                 .statement,
@@ -646,7 +646,7 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
                         refer(
                           'deserialize${_displayString(innerReturnType)}List',
                         ),
-                        refer('$_resultVar.data!.cast<Map<String,dynamic>>()')
+                        refer('$_resultVar.data!.cast<Map<String,dynamic>>()'),
                       ]),
                     ),
                   )
@@ -662,7 +662,6 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
                 mapperCode = refer(
                   '(dynamic i) => ${_displayString(innerReturnType)}.fromMap(i as $castType)',
                 );
-                break;
               case retrofit.Parser.JsonSerializable:
                 if (innerReturnType?.isNullable ?? false) {
                   mapperCode = refer(
@@ -673,12 +672,10 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
                     '(dynamic i) => ${_displayString(innerReturnType)}.fromJson(i as $castType)',
                   );
                 }
-                break;
               case retrofit.Parser.DartJsonMapper:
                 mapperCode = refer(
                   '(dynamic i) => JsonMapper.fromMap<${_displayString(innerReturnType)}>(i as $castType)!',
                 );
-                break;
               case retrofit.Parser.FlutterCompute:
                 throw Exception('Unreachable code');
             }
@@ -733,7 +730,6 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
                     .toList()
                 )
             ''');
-                break;
               case retrofit.Parser.JsonSerializable:
                 mapperCode = refer('''
             (k, dynamic v) =>
@@ -743,7 +739,6 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
                     .toList()
                 )
             ''');
-                break;
               case retrofit.Parser.DartJsonMapper:
                 mapperCode = refer('''
             (k, dynamic v) =>
@@ -753,7 +748,6 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
                     .toList()
                 )
             ''');
-                break;
               case retrofit.Parser.FlutterCompute:
                 log.warning('''
 Return types should not be a map when running `Parser.FlutterCompute`, as spawning an isolate per entry is extremely intensive.
@@ -766,7 +760,6 @@ You should create a new class to encapsulate the response.
                     await compute(deserialize${_displayString(type)}List,
                         (e.value as List).cast<Map<String, dynamic>>()))
             ''');
-                break;
             }
             if (future) {
               _wrapInTryCatch(
@@ -778,8 +771,8 @@ You should create a new class to encapsulate the response.
                       refer('Map.fromEntries').call([
                         refer('await Future.wait').call([
                           refer('$_resultVar.data!.entries.map')
-                              .call([mapperCode])
-                        ])
+                              .call([mapperCode]),
+                        ]),
                       ]),
                     )
                     .statement,
@@ -809,18 +802,15 @@ You should create a new class to encapsulate the response.
                 mapperCode = refer(
                   '(k, dynamic v) => MapEntry(k, ${_displayString(secondType)}.fromMap(v as Map<String, dynamic>))',
                 );
-                break;
               case retrofit.Parser.JsonSerializable:
                 mapperCode = refer(
                   '(k, dynamic v) => MapEntry(k, ${_displayString(secondType)}.fromJson(v as Map<String, dynamic>))',
                 );
 
-                break;
               case retrofit.Parser.DartJsonMapper:
                 mapperCode = refer(
                   '(k, dynamic v) => MapEntry(k, JsonMapper.fromMap<${_displayString(secondType)}>(v as Map<String, dynamic>)!)',
                 );
-                break;
               case retrofit.Parser.FlutterCompute:
                 log.warning('''
 Return types should not be a map when running `Parser.FlutterCompute`, as spawning an isolate per entry is extremely intensive.
@@ -831,7 +821,6 @@ You should create a new class to encapsulate the response.
                 (e) async => MapEntry(
                     e.key, await compute(deserialize${_displayString(secondType)}, e.value as Map<String, dynamic>))
             ''');
-                break;
             }
             if (future) {
               _wrapInTryCatch(
@@ -845,8 +834,8 @@ You should create a new class to encapsulate the response.
                         whenFalse: refer('Map.fromEntries').call([
                           refer('await Future.wait').call([
                             refer('$_resultVar.data!.entries.map')
-                                .call([mapperCode])
-                          ])
+                                .call([mapperCode]),
+                          ]),
                         ]),
                       ),
                     )
@@ -923,14 +912,19 @@ You should create a new class to encapsulate the response.
             )
             ..add(const Code('final $_valueVar = $_resultVar.data;'));
         } else if (_typeChecker(GeneratedMessage).isSuperTypeOf(returnType)) {
-          blocks.add(
-            declareFinal(_resultVar)
-                .assign(
-                    refer("await $_dioVar.fetch<List<int>>").call([options]))
-                .statement,
-          );
-          blocks.add(Code(
-              "final $_valueVar = await compute(${_displayString(returnType)}.fromBuffer, $_resultVar.data!);"));
+          blocks
+            ..add(
+              declareFinal(_resultVar)
+                  .assign(
+                    refer('await $_dioVar.fetch<List<int>>').call([options]),
+                  )
+                  .statement,
+            )
+            ..add(
+              Code(
+                'final $_valueVar = await compute(${_displayString(returnType)}.fromBuffer, $_resultVar.data!);',
+              ),
+            );
         } else {
           final fetchType = returnType.isNullable
               ? 'Map<String,dynamic>?'
@@ -950,7 +944,6 @@ You should create a new class to encapsulate the response.
               mapperCode = refer(
                 '${_displayString(returnType)}.fromMap($_resultVar.data!)',
               );
-              break;
             case retrofit.Parser.JsonSerializable:
               final genericArgumentFactories =
                   isGenericArgumentFactories(returnType);
@@ -975,9 +968,9 @@ You should create a new class to encapsulate the response.
               } else {
                 if (_isEnum(returnType) && !_hasFromJson(returnType)) {
                   mapperCode = refer(
-                    '${_displayString(returnType)}.values.firstWhere((e) => e.name == _result.data,'
-                    'orElse: () => throw ArgumentError('
-                    '\'${_displayString(returnType)} does not contain value \${_result.data}\','
+                    '${_displayString(returnType)}.values.firstWhere((e) => e.name == _result.data, '
+                    'orElse: () => throw ArgumentError( '
+                    "'${_displayString(returnType)} does not contain value \${_result.data}', "
                     '),'
                     ')',
                   );
@@ -987,17 +980,14 @@ You should create a new class to encapsulate the response.
                   );
                 }
               }
-              break;
             case retrofit.Parser.DartJsonMapper:
               mapperCode = refer(
                 'JsonMapper.fromMap<${_displayString(returnType)}>($_resultVar.data!)!',
               );
-              break;
             case retrofit.Parser.FlutterCompute:
               mapperCode = refer(
                 'await compute(deserialize${_displayString(returnType).replaceFirst('<', '').replaceFirst('>', '')}, $_resultVar.data!)',
               );
-              break;
           }
           _wrapInTryCatch(
             blocks,
@@ -1045,7 +1035,9 @@ You should create a new class to encapsulate the response.
         final obj = annotation.peek('genericArgumentFactories');
         // ignore: invalid_null_aware_operator
         genericArgumentFactories = obj?.boolValue ?? false;
-      } catch (_) {}
+      } on Object {
+        // nothing
+      }
     }
 
     return genericArgumentFactories ||
@@ -1055,19 +1047,29 @@ You should create a new class to encapsulate the response.
   /// Checks for a compatible fromJson signature for generic argument factories
   // TODO: But does the code work with multiple generic types?
   bool hasGenericArgumentFactoriesCompatibleSignature(DartType? dartType) {
-    if (dartType == null) return false;
+    if (dartType == null) {
+      return false;
+    }
     final element = dartType.element;
-    if (element is! InterfaceElement) return false;
+    if (element is! InterfaceElement) {
+      return false;
+    }
 
     final typeParameters = element.typeParameters;
-    if (typeParameters.isEmpty) return false;
+    if (typeParameters.isEmpty) {
+      return false;
+    }
 
     final constructors = element.constructors;
-    if (constructors.isEmpty) return false;
+    if (constructors.isEmpty) {
+      return false;
+    }
     final fromJson = constructors
         .firstWhereOrNull((constructor) => constructor.name == 'fromJson');
 
-    if (fromJson == null || fromJson.parameters.length == 1) return false;
+    if (fromJson == null || fromJson.parameters.length == 1) {
+      return false;
+    }
 
     final fromJsonArguments = fromJson.parameters;
 
@@ -1250,7 +1252,7 @@ You should create a new class to encapsulate the response.
 
       final composeArguments = <String, Expression>{
         _queryParamsVar: queryParams,
-        _dataVar: dataVar
+        _dataVar: dataVar,
       };
       if (cancelToken != null) {
         composeArguments[_cancelToken] = cancelToken;
@@ -1275,10 +1277,10 @@ You should create a new class to encapsulate the response.
               _baseUrlVar: refer('_combineBaseUrls').call([
                 refer(_dioVar).property('options').property('baseUrl'),
                 baseUrl,
-              ])
-            })
+              ]),
+            }),
       ], {}, [
-        type
+        type,
       ]);
     } else {
       hasCustomOptions = true;
@@ -1286,7 +1288,7 @@ You should create a new class to encapsulate the response.
         declareFinal('newOptions')
             .assign(
               refer('newRequestOptions')
-                  .call([refer(annoOptions.item1.displayName)]),
+                  .call([refer(annoOptions.element.displayName)]),
             )
             .statement,
       );
@@ -1384,7 +1386,7 @@ You should create a new class to encapsulate the response.
           ..returns = refer('String')
           ..requiredParameters =
               ListBuilder(<Parameter>[dioBaseUrlParam, baseUrlParam])
-          ..body = const Code(r'''
+          ..body = const Code('''
             if (baseUrl == null || baseUrl.trim().isEmpty) {
               return dioBaseUrl;
             }
@@ -1507,20 +1509,16 @@ if (T != dynamic &&
                   ? refer(p.displayName).nullSafeProperty('toJson').call([])
                   : refer(p.displayName).property('toJson').call([]);
             }
-            break;
           case retrofit.Parser.MapSerializable:
             value = p.type.nullabilitySuffix == NullabilitySuffix.question
                 ? refer(p.displayName).nullSafeProperty('toMap').call([])
                 : refer(p.displayName).property('toMap').call([]);
-            break;
           case retrofit.Parser.DartJsonMapper:
             value = refer(p.displayName);
-            break;
           case retrofit.Parser.FlutterCompute:
             value = refer(
               'await compute(serialize${_displayString(p.type)}, ${p.displayName})',
             );
-            break;
         }
       }
       return MapEntry(literalString(key, raw: true), value);
@@ -1550,20 +1548,16 @@ if (T != dynamic &&
             value = p.type.nullabilitySuffix == NullabilitySuffix.question
                 ? refer(displayName).nullSafeProperty('toJson').call([])
                 : refer(displayName).property('toJson').call([]);
-            break;
           case retrofit.Parser.MapSerializable:
             value = p.type.nullabilitySuffix == NullabilitySuffix.question
                 ? refer(displayName).nullSafeProperty('toMap').call([])
                 : refer(displayName).property('toMap').call([]);
-            break;
           case retrofit.Parser.DartJsonMapper:
             value = refer(displayName);
-            break;
           case retrofit.Parser.FlutterCompute:
             value = refer(
               'await compute(serialize${_displayString(p.type)}, ${p.displayName})',
             );
-            break;
         }
       }
 
@@ -1610,10 +1604,10 @@ if (T != dynamic &&
         _getMethodAnnotationByType(m, retrofit.PreventNullToAbsent);
 
     final annotation = _getAnnotation(m, retrofit.Body);
-    final bodyName = annotation?.item1;
+    final bodyName = annotation?.element;
     if (bodyName != null) {
       final nullToAbsent =
-          annotation!.item2.peek('nullToAbsent')?.boolValue ?? false;
+          annotation!.reader.peek('nullToAbsent')?.boolValue ?? false;
       final bodyTypeElement = bodyName.type.element;
       if (const TypeChecker.fromRuntime(Map)
           .isAssignableFromType(bodyName.type)) {
@@ -1627,7 +1621,7 @@ if (T != dynamic &&
             refer('$dataVar.addAll').call([
               refer(
                 "${bodyName.displayName}${m.type.nullabilitySuffix == NullabilitySuffix.question ? ' ?? <String,dynamic>{}' : ''}",
-              )
+              ),
             ]).statement,
           );
         if (preventNullToAbsent == null && nullToAbsent) {
@@ -1649,7 +1643,6 @@ if (T != dynamic &&
                   )
                   .statement,
             );
-            break;
           case retrofit.Parser.MapSerializable:
             blocks.add(
               declareFinal(dataVar)
@@ -1660,7 +1653,6 @@ if (T != dynamic &&
                   )
                   .statement,
             );
-            break;
           case retrofit.Parser.FlutterCompute:
             blocks.add(
               declareFinal(dataVar)
@@ -1671,16 +1663,18 @@ if (T != dynamic &&
                   )
                   .statement,
             );
-            break;
         }
       } else if (_typeChecker(GeneratedMessage).isSuperTypeOf(bodyName.type)) {
         if (bodyName.type.nullabilitySuffix != NullabilitySuffix.none) {
           log.warning(
-              "GeneratedMessage body ${_displayString(bodyName.type)} can not be nullable.");
+            'GeneratedMessage body ${_displayString(bodyName.type)} can not be nullable.',
+          );
         }
-        blocks.add(declareFinal(dataVar)
-            .assign(refer("${bodyName.displayName}.writeToBuffer()"))
-            .statement);
+        blocks.add(
+          declareFinal(dataVar)
+              .assign(refer('${bodyName.displayName}.writeToBuffer()'))
+              .statement,
+        );
       } else if (bodyTypeElement != null &&
           _typeChecker(File).isExactly(bodyTypeElement)) {
         blocks.add(
@@ -1695,7 +1689,8 @@ if (T != dynamic &&
       } else if (bodyName.type.element is ClassElement) {
         final ele = bodyName.type.element! as ClassElement;
         if (clientAnnotation.parser == retrofit.Parser.MapSerializable) {
-          final toMap = ele.lookUpMethod('toMap', ele.library);
+          final toMap =
+              ele.augmented.lookUpMethod(name: 'toMap', library: ele.library);
           if (toMap == null) {
             log.warning(
                 '${_displayString(bodyName.type)} must provide a `toMap()` method which return a Map.\n'
@@ -1717,7 +1712,7 @@ if (T != dynamic &&
                   [
                     refer(
                       '${bodyName.displayName}?.toMap() ?? <String,dynamic>{}',
-                    )
+                    ),
                   ],
                 ).statement,
               );
@@ -1768,7 +1763,7 @@ if (T != dynamic &&
                     NullabilitySuffix.question) {
                   blocks.add(
                     refer('$dataVar.addAll').call([
-                      refer('${bodyName.displayName}.toJson($toJsonCode)')
+                      refer('${bodyName.displayName}.toJson($toJsonCode)'),
                     ]).statement,
                   );
                 } else {
@@ -1776,11 +1771,10 @@ if (T != dynamic &&
                     refer('$dataVar.addAll').call([
                       refer(
                         '${bodyName.displayName}?.toJson($toJsonCode) ?? <String,dynamic>{}',
-                      )
+                      ),
                     ]).statement,
                   );
                 }
-                break;
               case retrofit.Parser.FlutterCompute:
                 if (bodyName.type.nullabilitySuffix !=
                     NullabilitySuffix.question) {
@@ -1788,7 +1782,7 @@ if (T != dynamic &&
                     refer('$dataVar.addAll').call([
                       refer(
                         'await compute(serialize${_displayString(bodyName.type)}, ${bodyName.displayName})',
-                      )
+                      ),
                     ]).statement,
                   );
                 } else {
@@ -1798,11 +1792,10 @@ if (T != dynamic &&
 ${bodyName.displayName} == null
                       ? <String, dynamic>{}
                       : await compute(serialize${_displayString(bodyName.type)}, ${bodyName.displayName})
-                  ''')
+                  '''),
                     ]).statement,
                   );
                 }
-                break;
               case retrofit.Parser.MapSerializable:
                 // Unreachable code
                 break;
@@ -1874,6 +1867,9 @@ ${bodyName.displayName} == null
         final contentType = r.peek('contentType')?.stringValue;
 
         if (isFileField) {
+          if (p.type.isNullable) {
+            blocks.add(Code('if (${p.displayName} != null){'));
+          }
           final fileNameValue = r.peek('fileName')?.stringValue;
           final fileName = fileNameValue != null
               ? literalString(fileNameValue)
@@ -1881,14 +1877,14 @@ ${bodyName.displayName} == null
                   .property('path.split(Platform.pathSeparator).last');
 
           final uploadFileInfo = refer('$MultipartFile.fromFileSync').call([
-            refer(p.displayName).property('path')
+            refer(p.displayName).property('path'),
           ], {
             'filename': fileName,
             if (contentType != null)
               'contentType':
                   refer('MediaType', 'package:http_parser/http_parser.dart')
                       .property('parse')
-                      .call([literal(contentType)])
+                      .call([literal(contentType)]),
           });
 
           final optionalFile = m.parameters
@@ -1898,7 +1894,7 @@ ${bodyName.displayName} == null
 
           final returnCode =
               refer(dataVar).property('files').property('add').call([
-            refer('MapEntry').newInstance([literal(fieldName), uploadFileInfo])
+            refer('MapEntry').newInstance([literal(fieldName), uploadFileInfo]),
           ]).statement;
           if (optionalFile) {
             final condition = refer(p.displayName).notEqualTo(literalNull).code;
@@ -1908,11 +1904,14 @@ ${bodyName.displayName} == null
                 condition,
                 const Code(') {'),
                 returnCode,
-                const Code('}')
+                const Code('}'),
               ],
             );
           } else {
             blocks.add(returnCode);
+          }
+          if (p.type.isNullable) {
+            blocks.add(Code('}'));
           }
         } else if (_displayString(p.type) == 'List<int>') {
           final optionalFile = m.parameters
@@ -1933,7 +1932,7 @@ ${bodyName.displayName} == null
                 filename:${literal(fileName)},
                     $conType
                     ))
-                  ''')
+                  '''),
           ]).statement;
           if (optionalFile) {
             final condition = refer(p.displayName).notEqualTo(literalNull).code;
@@ -1943,7 +1942,7 @@ ${bodyName.displayName} == null
                 condition,
                 const Code(') {'),
                 returnCode,
-                const Code('}')
+                const Code('}'),
               ],
             );
           } else {
@@ -1967,7 +1966,7 @@ ${bodyName.displayName} == null
                     filename:${literal(fileName)},
                     $conType
                     )))
-                  ''')
+                  '''),
               ]).statement,
             );
           } else if (_isBasicType(innerType) ||
@@ -2005,7 +2004,7 @@ ${bodyName.displayName} == null
                     filename: i.path.split(Platform.pathSeparator).last,
                     $conType
                     )))
-                  ''')
+                  '''),
               ]).statement,
             );
             if (p.type.isNullable) {
@@ -2022,7 +2021,7 @@ ${bodyName.displayName} == null
                   ${p.displayName}.map((i) => MapEntry(
                 '$fieldName',
                 i))
-                  ''')
+                  '''),
               ]).statement,
             );
             if (p.type.isNullable) {
@@ -2033,13 +2032,12 @@ ${bodyName.displayName} == null
             if (_missingToJson(ele)) {
               if (_isDateTime(p.type)) {
                 final expr = [
-                  p.type.nullabilitySuffix == NullabilitySuffix.question
-                      ? refer(p.displayName)
-                          .nullSafeProperty('toIso8601String')
-                          .call([])
-                      : refer(p.displayName)
-                          .property('toIso8601String')
-                          .call([])
+                  if (p.type.nullabilitySuffix == NullabilitySuffix.question)
+                    refer(p.displayName)
+                        .nullSafeProperty('toIso8601String')
+                        .call([])
+                  else
+                    refer(p.displayName).property('toIso8601String').call([]),
                 ];
                 refer(dataVar).property('fields').property('add').call(expr);
               } else {
@@ -2050,8 +2048,8 @@ ${bodyName.displayName} == null
                 refer(dataVar).property('fields').property('add').call([
                   refer('MapEntry').newInstance([
                     literal(fieldName),
-                    refer('jsonEncode(${p.displayName})')
-                  ])
+                    refer('jsonEncode(${p.displayName})'),
+                  ]),
                 ]).statement,
               );
             }
@@ -2071,8 +2069,8 @@ ${bodyName.displayName} == null
                 else if (_isEnum(p.type))
                   refer(p.displayName).property('name')
                 else
-                  refer(p.displayName).property('toString').call([])
-              ])
+                  refer(p.displayName).property('toString').call([]),
+              ]),
             ]).statement,
           );
           if (p.type.nullabilitySuffix == NullabilitySuffix.question) {
@@ -2084,7 +2082,7 @@ ${bodyName.displayName} == null
             refer(dataVar).property('fields').property('add').call([
               refer('MapEntry').newInstance(
                 [literal(fieldName), refer('jsonEncode(${p.displayName})')],
-              )
+              ),
             ]).statement,
           );
         } else if (p.type.element is ClassElement) {
@@ -2092,11 +2090,12 @@ ${bodyName.displayName} == null
           if (_missingToJson(ele)) {
             if (_isDateTime(p.type)) {
               final expr = [
-                p.type.nullabilitySuffix == NullabilitySuffix.question
-                    ? refer(p.displayName)
-                        .nullSafeProperty('toIso8601String')
-                        .call([])
-                    : refer(p.displayName).property('toIso8601String').call([])
+                if (p.type.nullabilitySuffix == NullabilitySuffix.question)
+                  refer(p.displayName)
+                      .nullSafeProperty('toIso8601String')
+                      .call([])
+                else
+                  refer(p.displayName).property('toIso8601String').call([]),
               ];
               refer(dataVar).property('fields').property('add').call(expr);
             } else {
@@ -2107,12 +2106,12 @@ ${bodyName.displayName} == null
               final uploadFileInfo = refer('$MultipartFile.fromString').call([
                 refer(
                   "jsonEncode(${p.displayName}${p.type.nullabilitySuffix == NullabilitySuffix.question ? ' ?? <String,dynamic>{}' : ''})",
-                )
+                ),
               ], {
                 'contentType':
                     refer('MediaType', 'package:http_parser/http_parser.dart')
                         .property('parse')
-                        .call([literal(contentType)])
+                        .call([literal(contentType)]),
               });
 
               final optionalFile = m.parameters
@@ -2123,7 +2122,7 @@ ${bodyName.displayName} == null
               final returnCode =
                   refer(dataVar).property('files').property('add').call([
                 refer('MapEntry')
-                    .newInstance([literal(fieldName), uploadFileInfo])
+                    .newInstance([literal(fieldName), uploadFileInfo]),
               ]).statement;
               if (optionalFile) {
                 final condition =
@@ -2134,7 +2133,7 @@ ${bodyName.displayName} == null
                     condition,
                     const Code(') {'),
                     returnCode,
-                    const Code('}')
+                    const Code('}'),
                   ],
                 );
               } else {
@@ -2147,8 +2146,8 @@ ${bodyName.displayName} == null
                     literal(fieldName),
                     refer(
                       "jsonEncode(${p.displayName}${p.type.nullabilitySuffix == NullabilitySuffix.question ? ' ?? <String,dynamic>{}' : ''})",
-                    )
-                  ])
+                    ),
+                  ]),
                 ]).statement,
               );
             }
@@ -2157,7 +2156,7 @@ ${bodyName.displayName} == null
           blocks.add(
             refer(dataVar).property('fields').property('add').call([
               refer('MapEntry')
-                  .newInstance([literal(fieldName), refer(p.displayName)])
+                  .newInstance([literal(fieldName), refer(p.displayName)]),
             ]).statement,
           );
         }
@@ -2166,7 +2165,7 @@ ${bodyName.displayName} == null
     }
 
     /// There is no body
-    if (globalOptions.emptyRequestBody == true) {
+    if (globalOptions.emptyRequestBody ?? false) {
       blocks.add(
         declareFinal(dataVar)
             .assign(literalMap({}, refer('String'), refer('dynamic')))
@@ -2190,13 +2189,13 @@ ${bodyName.displayName} == null
               dynamic val;
               if (v == null) {
                 val = null;
-              } else if (v.type?.isDartCoreBool == true) {
+              } else if (v.type?.isDartCoreBool ?? false) {
                 val = v.toBoolValue();
-              } else if (v.type?.isDartCoreString == true) {
+              } else if (v.type?.isDartCoreString ?? false) {
                 val = v.toStringValue();
-              } else if (v.type?.isDartCoreDouble == true) {
+              } else if (v.type?.isDartCoreDouble ?? false) {
                 val = v.toDoubleValue();
-              } else if (v.type?.isDartCoreInt == true) {
+              } else if (v.type?.isDartCoreInt ?? false) {
                 val = v.toIntValue();
               } else {
                 val = v.toStringValue();
@@ -2225,12 +2224,15 @@ ${bodyName.displayName} == null
 
     if (returnType != null &&
         _typeChecker(GeneratedMessage).isAssignableFromType(returnType)) {
-      headers.removeWhere(
-          (key, value) => "accept".toLowerCase() == key.toLowerCase());
-      headers.addAll({
-        "accept": literal(
-            "application/x-protobuf; \${${_displayString(returnType)}.getDefault().info_.qualifiedMessageName == \"\" ? \"\" :\"messageType=\${${_displayString(returnType)}.getDefault().info_.qualifiedMessageName}\"}")
-      });
+      headers
+        ..removeWhere(
+          (key, value) => 'accept'.toLowerCase() == key.toLowerCase(),
+        )
+        ..addAll({
+          'accept': literal(
+            'application/x-protobuf; \${${_displayString(returnType)}.getDefault().info_.qualifiedMessageName == "" ? "" :"messageType=\${${_displayString(returnType)}.getDefault().info_.qualifiedMessageName}"}',
+          ),
+        });
     }
 
     return headers;
@@ -2265,7 +2267,7 @@ ${bodyName.displayName} == null
         if (noStore ?? false) 'no-store' else '',
         if (noTransform ?? false) 'no-transform' else '',
         if (onlyIfCached ?? false) 'only-if-cached' else '',
-        ...otherResult
+        ...otherResult,
       ];
 
       final value = values.where((element) => element != '').join(', ');
@@ -2273,6 +2275,68 @@ ${bodyName.displayName} == null
       result.putIfAbsent(HttpHeaders.cacheControlHeader, () => literal(value));
     }
     return result;
+  }
+
+  Object? _getFieldValue(ConstantReader? value) {
+    if (value?.isBool ?? false) return value?.boolValue;
+    if (value?.isDouble ?? false) return value?.doubleValue;
+    if (value?.isInt ?? false) return value?.intValue;
+    if (value?.isString ?? false) return value?.stringValue;
+    if (value?.objectValue.isEnum ?? false) {
+      return value?.objectValue.variable?.displayName;
+    }
+    if (value?.isList ?? false) {
+      return value?.listValue
+          .map((item) => _getFieldValue(ConstantReader(item)))
+          .toList();
+    }
+    if (value?.isMap ?? false) {
+      final mapValue = value?.mapValue.map((key, val) {
+        return MapEntry(
+          _getFieldValue(ConstantReader(key)),
+          _getFieldValue(ConstantReader(val)),
+        );
+      });
+      return mapValue;
+    }
+    if (value?.isSet ?? false) {
+      return value?.setValue.map((item) {
+        return _getFieldValue(ConstantReader(item));
+      }).toSet();
+    }
+    if (value?.objectValue.type != null) {
+      final fields = <String, Object?>{};
+      final type = value!.objectValue.type;
+      if (type is InterfaceType) {
+        for (var field in type.element.fields) {
+          if (!field.isStatic) {
+            final fieldValue = value.peek(field.name);
+            fields[field.name] = _getFieldValue(fieldValue);
+          }
+        }
+      }
+      return fields;
+    }
+    return null;
+  }
+
+  Map<String, Object> _getMapFromTypedExtras(MethodElement m) {
+    final annotations = _getMethodAnnotations(m, retrofit.TypedExtras);
+    final allTypedExtras = <String, Object>{};
+
+    for (final annotation in annotations) {
+      final fields = annotation.objectValue.type?.element?.children
+          .whereType<FieldElement>();
+      for (final field in fields ?? <FieldElement>[]) {
+        final value = annotation.peek(field.name);
+        final fieldValue = _getFieldValue(value);
+        if (fieldValue != null) {
+          allTypedExtras[field.name] = fieldValue;
+        }
+      }
+    }
+
+    return allTypedExtras;
   }
 
   void _generateExtra(
@@ -2310,7 +2374,10 @@ ${bodyName.displayName} == null
                       ),
                     ),
                   )
-                  .fold<Map<String, Object>>({}, (p, e) => p..addAll(e ?? {})),
+                  .fold<Map<String, Object>>({}, (p, e) {
+                return p..addAll(e ?? {});
+              })
+                ..addAll(_getMapFromTypedExtras(m)),
               refer('String'),
               refer('dynamic'),
             ),
@@ -2335,20 +2402,16 @@ ${bodyName.displayName} == null
             value = p.type.nullabilitySuffix == NullabilitySuffix.question
                 ? refer(displayName).nullSafeProperty('toJson').call([])
                 : refer(displayName).property('toJson').call([]);
-            break;
           case retrofit.Parser.MapSerializable:
             value = p.type.nullabilitySuffix == NullabilitySuffix.question
                 ? refer(displayName).nullSafeProperty('toMap').call([])
                 : refer(displayName).property('toMap').call([]);
-            break;
           case retrofit.Parser.DartJsonMapper:
             value = refer(displayName);
-            break;
           case retrofit.Parser.FlutterCompute:
             value = refer(
               'await compute(serialize${_displayString(p.type)}, ${p.displayName})',
             );
-            break;
         }
       }
 
@@ -2368,7 +2431,8 @@ ${bodyName.displayName} == null
     switch (clientAnnotation.parser) {
       case retrofit.Parser.JsonSerializable:
       case retrofit.Parser.DartJsonMapper:
-        final toJson = ele.lookUpMethod('toJson', ele.library);
+        final toJson =
+            ele.augmented.lookUpMethod(name: 'toJson', library: ele.library);
         return toJson == null;
       case retrofit.Parser.MapSerializable:
       case retrofit.Parser.FlutterCompute:
@@ -2394,7 +2458,11 @@ ${bodyName.displayName} == null
   }
 
   void _wrapInTryCatch(
-      List<Code> blocks, Expression options, DartType? returnType, Code child) {
+    List<Code> blocks,
+    Expression options,
+    DartType? returnType,
+    Code child,
+  ) {
     blocks.addAll(
       [
         declareVar(
@@ -2402,12 +2470,12 @@ ${bodyName.displayName} == null
           type: refer(_displayString(returnType, withNullability: true)),
           late: true,
         ).statement,
-        Code('try {'),
+        const Code('try {'),
         child,
-        Code('} on Object catch (e, s) {'),
-        Code('$_errorLoggerVar?.logError(e, s, $_optionsVar);'),
-        Code('rethrow;'),
-        Code('}'),
+        const Code('} on Object catch (e, s) {'),
+        const Code('$_errorLoggerVar?.logError(e, s, $_optionsVar);'),
+        const Code('rethrow;'),
+        const Code('}'),
       ],
     );
   }
@@ -2555,16 +2623,36 @@ extension DartTypeStreamAnnotation on DartType {
 
 String _displayString(DartType? e, {bool withNullability = false}) {
   try {
-    return e!.getDisplayString(withNullability: withNullability);
+    if (!withNullability) {
+      return e!.toStringNonNullable();
+    } else {
+      return e!.getDisplayString();
+    }
   } on TypeError {
-    return e!.getDisplayString(withNullability: withNullability);
-  } on Object {
-    rethrow;
+    if (!withNullability) {
+      return e!.toStringNonNullable();
+    } else {
+      return e!.getDisplayString();
+    }
   }
 }
 
 extension DartTypeExt on DartType {
   bool get isNullable => nullabilitySuffix == NullabilitySuffix.question;
+
+  String toStringNonNullable() {
+    final val = getDisplayString();
+    if (val.endsWith('?')) {
+      return val.substring(0, val.length - 1);
+    }
+    return val;
+  }
+}
+
+extension DartObjectX on DartObject? {
+  bool get isEnum {
+    return this?.type?.element?.kind.name == 'ENUM';
+  }
 }
 
 extension ReferenceExt on Reference {
