@@ -1,3 +1,3 @@
 export 'dio.dart';
-export 'http.dart';
 export 'error_logger.dart';
+export 'http.dart';
